拆分几何基类
