+++ src/manimgeo/components/base/__init__.py
@@ -1,0 +1,6 @@
+from .base_geometry import BaseGeometry
+from .base_adapter import GeometryAdapter
+
+# 重建
+BaseGeometry.model_rebuild()
+GeometryAdapter.model_rebuild()
